Fix strict release build with gcc 14
As of at least gcc 14.2.1, Wmaybe-uninitialized has a false positive for the
_size field of StaticVector (which is guaranteed by C++ to always be
initialized) at optimization level 2 or higher.

The gcc man page alludes to unreliability of this warning (although somehow
associated with sanitizers), and many gcc bugs have been filed over the past
several years about it, so just disable it.  This mistake would be caught in
several other ways anyway (like Wuninitialized in GCC or hicpp-member-init in
clang-tidy).